+++ lib/monitor.js
@@ -118,7 +118,6 @@ class Monitor {
                 let stream = undefined;
                 let ctorOpts;
 
-                //TODO: Check for "special reporters" here and break early
                 if (typeof spec === 'string') {
                     ctorOpts = {
                         module: spec
