+++ lib/schema.js
@@ -19,7 +19,7 @@ exports.monitor = Joi.object().keys({
                 module: Joi.string().required(),
                 name: Joi.string(),
                 args: Joi.array().default([])
-            }),
+            }).required(),
             start: Joi.array().default([])
         })
     )).default({}),
