+++ test/fixtures/reporter.js
